--- src/main/java/com/kodilla/ecommercee/order/OrderController.java
@@ -1,37 +1,0 @@
-package com.kodilla.ecommercee.order;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@RestController
-@RequestMapping("/v1/ecommercee/order")
-public class OrderController {
-    @RequestMapping(method = RequestMethod.GET, value = "getOrderList")
-    public List<OrderDto> getOrderList(){
-        return new ArrayList<>();
-    }
-
-    @RequestMapping(method = RequestMethod.POST, value = "createOrder")
-    public void createOrder(OrderDto orderDto){
-
-    }
-
-    @RequestMapping(method = RequestMethod.GET, value = "getOrder")
-    public OrderDto getOrder(Long orderId){
-        return new OrderDto(orderId,10L,15L);
-    }
-
-    @RequestMapping(method = RequestMethod.PUT, value = "updateOrder")
-    public OrderDto updateOrder(OrderDto orderDto){
-        return new OrderDto(1L, 13L, 17L);
-    }
-
-    @RequestMapping(method = RequestMethod.DELETE, value = "deleteOrder")
-    public void deleteOrder(Long orderId){
-
-    }
-}
